+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Documents changes that result in:
 ## Unreleased
 
 - Add a deployment-time configurable limit on the whole balance of UserDeposit
+- Deployment script's `service` command takes an additional option `--user-deposit-whole-limit`
 
 ## [0.13.0](https://github.com/raiden-network/raiden-contracts/releases/tag/v0.13.0) - 2019-03-04